chore(sdk/elixir): cleanup tests
